+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **breaking** ARP now returns an error on non-supported operation codes
 - **breaking** Change `ipv4_options` and `ipv4_options_mut` from Option to empty slices if no options are present
 - Improve IPv6 `new()` performance by reducing buffer accesses
+- Properly handle atomic IPv6 fragment extensions
 
 ### Fixed
 - Return an error on a hop by hop IPv6 extension that is not the first extension